0.964  2022/03/10  SNOOPYJC  Various fixes
Don't mark variables as initialized if there is a conditional at the end of the statement,
don't generate bad return statment if last expression of sub is an unshift,
handle @1234 as a variable but give a warning,
handle assignment to args and add a warning,
remove re.G on a regex in a non-normal place,
fix return {}, add -l flag and default to 98 (was 88),
strip escape chars where not ignored by python e.g. \[,
handle method call with {} in place of argument list,
initialize vars that are conditionally set in an eval block,
handle TYPE on my/our declarations (and ignore it),
in print mysub() - mysub is NOT a file handle, $^X was calling sys.executable() instead of referencing sys.executable properly,
fix eval: If no error occurs, eval sets  to the empty string (we were setting it to None),
calling a local sub in an expression was incorrectly setting all subsequent terms to be in list context,
regex (qr) variable not being handled properly when passed as arg and used in match,
generate proper code for my/local/our hash or array initialization when extra parens are used around variable name.
Library: Fix _init_package for package with dotted name, don't raise exceptions in -C, -A, -M, fix Array __setitem__ with slice